In process of fixing hash table class.

+++ backend/code/arduino/embedded-code/include/config.hpp
@@ -3,8 +3,8 @@
 
 namespace Config
 {
-    constexpr uint8_t RED_LED_PIN = 11;
-    constexpr uint8_t BASE_ARRAY_SIZE = 10;
-    constexpr uint8_t DEFAULT_ARRAY_ADDITION = 5;       // default amount to increase an array size by
-    constexpr uint8_t BASE_HASH_TABLE_SIZE = 10;
+    constexpr int RED_LED_PIN = 11;
+    constexpr size_t BASE_ARRAY_SIZE = 10;
+    constexpr size_t DEFAULT_ARRAY_ADDITION = 5;       // default amount to increase an array size by
+    constexpr size_t BASE_HASH_TABLE_SIZE = 10;
 }
